+++ frontend/package.json
@@ -11,6 +11,7 @@
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.2",
     "react-hot-toast": "^2.3.0",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.1",
     "react-spinners": "^0.13.4",
     "web-vitals": "^2.1.4"
